Attempted to Create Non VR Scene
Tried to create a scene for non VR midi logging but this didn't work.
So it's basically back to the same state. There was however some
refactoring done the main Game Control Design.
GameController now it's own object decoupled from the Theremin.
MidiLogging is also its own component so it is no longer handled in
the Hand Controllers.