Next Note Directcion Arrow Signifier
This update adds a new arrow signifier to indicate the direction of the next note.
This is intended to give the user some indication which way they should prepare
to move.
There was also a refactor in the Note Ctrlr.
Now the Note Message contains a flag as to the note type (Start, End, Last)
The Note Ctrlr also notifies all subscribers when score is loaded so that this
controls the starting messages.